requirements.txt
Django==2.1.1
django-bootstrap4==0.0.7
django-model-utils
psycopg2==2.7.5
pytz==2018.5
six==1.11.0
